Match new game receipts to correct tab
Setting up a new game requires both new game receipt and the new game
event. Hold whichever come first in a pendingStart field in the store
and wait for the second message. When both pieces of data have been
received then populate the screen accordingly.

A pending tab is shown on the UI while waiting for the second piece of
data to arrive.

Accepting challenges from other opponents uses the original object
because there is no receipt in this case