Detect MetaMask account change
Poll MetaMask every 2 seconds and check if the accounts changed. Yes,
this is horrible, but right now it's the recommended mechanism. When the
account changes raise an event from Connect4Web3 and send actions to
clear the account data in the UI and reload the active games.
Re-register web3 events in Connect4Web3 also.