Refactor gamePlay reducers
Way too much crap code in there. Split out logic to handle specific
games in to a series of separate reducers. These are held in a map keyed
on action.type. gamePlay will try to find a reducer from the
gameHandlers which will handle the action. If found the gameHandler will
take care of the action, othewise the usual switch logic applies.