Fix race condition with pending moves
The pendingMove action enables the spinner on the UI to indicate that
the UI is waiting for a move to be included by the Ethereum network. When
the block is mined an event is raised which causes the UI to accept the
move and turn off the pending state.

When these two event happen in reverse order the move is accepted but
the UI stays in a pending state.

This fix prevents the action thunk from dispatching pendingMove() if the
UI has already seen the move accepted event. It detects this scenario by
checking that the player is the next one to move. If true, the next move
event cannot have been read yet.

Similarly for resigning and claiming a win. pendingMove() is only
dispatched if the game is not over yet.